refactor(auth): separate auth and jwt service
