+++ src/auth/guards/super-user.guard.ts
@@ -1,6 +1,6 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common'
 import { ModuleRef } from '@nestjs/core'
-import { AuthService } from '../auth.service'
+import { AuthService } from '../services/auth.service'
 
 @Injectable()
 export class SuperUserGuard implements CanActivate {
